% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datas.R
\docType{data}
\name{arrivals}
\alias{arrivals}
\title{Arrivals}
\format{
A \code{data.frame} with 1000 rows and 4 columns:
\itemize{
\item \code{gender} is a factor with two levels : \emph{female} and \emph{male}.
\item \code{time} is the person arrival time, in minutes.
\item \code{weight} is the person weight, as disclosed by her/himself, in kilograms.
\item \code{age} is the person age, as disclosed by her/himself, in years.
}
}
\usage{
arrivals
}
\description{
Arrivals data of men/women in a shop, starting at an arbitrary time
}
\author{
Philippe Hauchamps \href{mailto:philippe.hauchamps@student.uclouvain.be}{philippe.hauchamps@student.uclouvain.be}
}
\keyword{datasets}
