% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datas.R
\docType{data}
\name{diamonds2}
\alias{diamonds2}
\title{diaomonds2}
\format{
A \code{data.frame} with 400 rows and 10 columns:
\itemize{
\item \code{price} price in US dollars ($326–$18,823)
\item \code{carat} weight of the diamond (0.2–5.01)
\item \code{cut} quality of the cut (Fair, Good, Very Good, Premium, Ideal)
\item \code{color} diamond colour, from D (best) to J (worst)
\item \code{clarity} a measurement of how clear the diamond is (I1 (worst), SI2, SI1, VS2, VS1, VVS2, VVS1, IF (best))
\item \code{x} length in mm (0–10.74)
\item \code{y} width in mm (0–58.9)
\item \code{z} depth in mm (0–31.8)
\item \code{depth} total depth percentage = z / mean(x, y) = 2.z / (x + y) (43--79)
\item \code{table} width of top of diamond relative to widest point (43–95)
}
}
\source{
extrait du package ggplot2
}
\usage{
diamonds2
}
\description{
A dataset containing the prices and other attributes of almost 54,000 diamonds.
The variables are as follows:
}
\references{
\url{https://github.com/tidyverse/ggplot2/blob/master/man/diamonds.Rd}
}
\author{
Trudel \href{mailto:balabla@mail.org}{balabla@mail.org}
}
\keyword{datasets}
