% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{iris}
\alias{iris}
\title{DataPack}
\format{
A data frame with 150 rows and 5 colums :
\itemize{
\item sep is ...
\item pet is ...
}
}
\usage{
iris
}
\description{
This is my second data set, it's a copy of the famous IRIS data set
}
\author{
Myself \href{mailto:myselfandI@hello.com}{myselfandI@hello.com}
}
\keyword{datasets}
