% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{normality.df}
\alias{normality.df}
\title{normality.df}
\usage{
normality.df(df, output = c("none", "figure", "message", "all"))
}
\arguments{
\item{df}{the data frame to be tested}

\item{output}{if "none", nothing is displayed, if "figure", a figure is displayed, if "message" a message is displayed, and  if "all", a message and a figure are displayed. Try and see !}
}
\value{
retourne une liste de deux elements a savoir:
\itemize{
\item \code{pval} p-value of the normality test
\item \code{result} text containing the results from the normality test
}
}
\description{
\code{normality.df} permet de faire le test de normalité de toutes les variables
numeriques d'un dataframe.
}
\examples{

# Example 1:
#---------------------------------------------
# apply normality.df to a data.frame
#---------------------------------------------
normality.df(diamonds2, output = "all")

}
