% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityTest-package.R
\docType{package}
\name{NormalityTest-package}
\alias{NormalityTest}
\alias{NormalityTest-package}
\title{NormalityTest: Test the normality of your data}
\description{
This package is build to test the normality of a data vector or a data frame via shapiro test and visual analysis.
}
\author{
\strong{Maintainer}: Vincent Guiot \email{vincent.guiot@student.uclouvain.be}

}
\keyword{internal}
