% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityTest.R
\name{normality.df}
\alias{normality.df}
\title{Function normality.df}
\usage{
normality.df(df, output = c("none", "figure", "message", "all"))
}
\arguments{
\item{df}{the data frame to be tested}

\item{output}{if \emph{"none"}, nothing is displayed, if \emph{"figure"}, a figure is
displayed, if \emph{"message"} a message is displayed, and  if \emph{"all"}, a message
and a figure are displayed}
}
\value{
a list with the following members :
\itemize{
\item \code{pvalues} : p-values of the normality test for the data frame
\item \code{results} : text containing the results from the normality test for the
data frame
}
}
\description{
Tests the normality of a all numerical variables of a given
data.frame using Shapiro-Wilk's test.
}
\examples{
#---------------------------------------------
# apply normality.df to a data.frame
#---------------------------------------------
normality.df(arrivals, output = "all")
}
