% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityTest-package.R
\docType{package}
\name{NormalityTest-package}
\alias{NormalityTest}
\alias{NormalityTest-package}
\title{NormalityTest: Utility package for testing normality}
\description{
This package provides a set of utility functions to test data normality. The data can be standalone or part of data.frames.
}
\author{
\strong{Maintainer}: Philippe Hauchamps \email{philippe.hauchamps@student.uclouvain.be} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
