% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{test.normality}
\alias{test.normality}
\title{test.normality}
\usage{
test.normality(
  x,
  output = c("none", "figure", "message", "all"),
  var.name = deparse(substitute(x))
)
}
\arguments{
\item{x}{tested variable}

\item{output}{if "none", nothing is displayed, if "figure", a figure is displayed, if "message" a message is displayed, and  if "all", a message and a figure are displayed. Try and see !}

\item{var.name}{the name to be displayed. The default value is the varname of x. Try and see !}
}
\value{
retourne une liste de deux elements a savoir:
\itemize{
\item \code{pval} p-value of the normality test
\item \code{result} text containing the results from the normality test
}
}
\description{
\code{test.normality} permet de faire le test de normalité des variables
numeriques d'un dataframe.
}
